Check that file exists in test case

+++ mclearn/tests/test_preprocessing.py
@@ -83,5 +83,6 @@ class TestHDF:
 
     @classmethod
     def teardown_class(cls):
-        os.remove(cls.hdf_path)
+        if os.path.isfile(self.hdf_path):
+            os.remove(cls.hdf_path)
